# openapi/trust-signals.yaml
openapi: 3.1.0
info:
  title: Trust Signals API
  version: 1.0.0
  description: |
    API for ingesting session events, computing trust signals, and managing
    step-up verification challenges. Server-to-server only — never expose
    API keys to browser code.

servers:
  - url: http://localhost:8080
    description: Local development

security:
  - ApiKeyAuth: []

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Server-side API key. Never expose to browsers.

  headers:
    X-RateLimit-Limit:
      description: Maximum requests allowed in the current window
      schema:
        type: integer
        example: 100
    X-RateLimit-Remaining:
      description: Requests remaining in the current window
      schema:
        type: integer
        example: 87
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit resets
      schema:
        type: integer
        example: 1712000000
    Retry-After:
      description: Seconds to wait before retrying (on 429)
      schema:
        type: integer
        example: 30

  schemas:
    SessionEvent:
      type: object
      required:
        - event_id
        - account_id
        - session_id
        - timestamp
        - client
        - device
        - network
      properties:
        event_id:
          type: string
          description: Globally unique event identifier for idempotency
          example: evt_abc123def456
        account_id:
          type: string
          description: Publisher-defined stable user identifier (opaque string)
          example: user_42
        session_id:
          type: string
          description: Session identifier from the publisher
          example: sess_xyz789
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of the event
          example: "2026-04-01T10:15:00Z"
        client:
          type: object
          required:
            - surface
          properties:
            surface:
              type: string
              enum:
                - publisher_backend
                - batch_fixture
                - manual_test
              description: Origin of the event
            sdk_version:
              type: string
              example: "1.2.0"
        device:
          type: object
          required:
            - fingerprint
          properties:
            fingerprint:
              type: string
              description: Device fingerprint hash
              example: fp_aabbccdd
            screen:
              type: string
              description: Screen resolution
              example: "1920x1080"
            timezone:
              type: string
              description: IANA timezone
              example: America/New_York
            locale:
              type: string
              description: Browser/device locale
              example: en-US
        network:
          type: object
          required:
            - ip
          properties:
            ip:
              type: string
              description: Client IP address (v4 or v6)
              example: "203.0.113.42"

    IngestRequest:
      type: object
      required:
        - events
      properties:
        events:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/SessionEvent'

    IngestResponse:
      type: object
      properties:
        accepted:
          type: integer
          description: Number of new events accepted
          example: 5
        duplicates:
          type: integer
          description: Number of duplicate events skipped
          example: 0
        errors:
          type: array
          items:
            type: object
            properties:
              event_id:
                type: string
              error:
                type: string
          description: Per-event validation errors (if any)

    AccountSignals:
      type: object
      properties:
        account_id:
          type: string
          example: user_42
        event_count:
          type: integer
          description: Total events received for this account
          example: 47
        distinct_sessions:
          type: integer
          description: Unique session_id count
          example: 3
        distinct_fingerprints:
          type: integer
          description: Unique device fingerprint count
          example: 2
        distinct_screens:
          type: integer
          description: Unique screen resolution count
          example: 2
        distinct_ip_prefixes:
          type: integer
          description: Unique /24 IP prefix count
          example: 4
        first_seen:
          type: string
          format: date-time
          example: "2026-03-15T08:00:00Z"
        last_seen:
          type: string
          format: date-time
          example: "2026-04-01T10:15:00Z"

    RiskResponse:
      type: object
      properties:
        account_id:
          type: string
          example: user_42
        risk_score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Composite risk score
          example: 0.72
        breakdown:
          type: object
          description: Individual factor contributions
          properties:
            fingerprint_factor:
              type: number
              example: 0.30
            ip_factor:
              type: number
              example: 0.25
            session_velocity_factor:
              type: number
              example: 0.17
          additionalProperties:
            type: number
        recommendation:
          type: string
          enum:
            - allow
            - challenge
            - block
          description: Suggested action based on risk score
          example: challenge

    ChallengeRequest:
      type: object
      required:
        - account_id
        - session_id
      properties:
        account_id:
          type: string
          example: user_42
        session_id:
          type: string
          example: sess_xyz789
        email:
          type: string
          format: email
          description: Email to send OTP (optional for testing)
          example: user@example.com

    ChallengeResponse:
      type: object
      properties:
        challenge_id:
          type: string
          description: Unique identifier for this challenge
          example: chal_abc123
        expires_at:
          type: string
          format: date-time
          description: When the challenge expires
          example: "2026-04-01T10:20:00Z"
        attempts_remaining:
          type: integer
          description: Verification attempts remaining
          example: 3
        code:
          type: string
          description: |
            The OTP code (only included in test mode for debugging).
            In production, this would be sent via email.
          example: "847291"

    VerifyRequest:
      type: object
      required:
        - challenge_id
        - code
      properties:
        challenge_id:
          type: string
          example: chal_abc123
        code:
          type: string
          pattern: "^[0-9]{6}$"
          description: 6-digit OTP code
          example: "847291"

    VerifyResponse:
      type: object
      properties:
        verified:
          type: boolean
          description: Whether the code was correct
          example: true
        account_id:
          type: string
          example: user_42
        session_id:
          type: string
          example: sess_xyz789

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error code
          example: invalid_request
        message:
          type: string
          description: Human-readable error message
          example: "Missing required field: account_id"
        details:
          type: object
          additionalProperties: true
          description: Additional error context

paths:
  /health:
    get:
      summary: Health check
      operationId: healthCheck
      security: []
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time

  /v1/events:
    post:
      summary: Ingest session events
      operationId: ingestEvents
      description: |
        Accept a batch of session events. Duplicate event_id values are
        silently ignored (idempotent). Use the Idempotency-Key header for
        request-level deduplication.
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
          description: Optional key for request-level idempotency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
            example:
              events:
                - event_id: evt_abc123
                  account_id: user_42
                  session_id: sess_xyz
                  timestamp: "2026-04-01T10:15:00Z"
                  client:
                    surface: publisher_backend
                    sdk_version: "1.2.0"
                  device:
                    fingerprint: fp_aabbcc
                    screen: "1920x1080"
                    timezone: America/New_York
                    locale: en-US
                  network:
                    ip: "203.0.113.42"
      responses:
        '200':
          description: Events processed
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              $ref: '#/components/headers/Retry-After'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: rate_limit_exceeded
                message: Too many requests. Retry after 30 seconds.

  /v1/accounts/{account_id}/signals:
    get:
      summary: Get aggregated signals for an account
      operationId: getAccountSignals
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
          example: user_42
      responses:
        '200':
          description: Aggregated signals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSignals'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/accounts/{account_id}/risk:
    get:
      summary: Get risk score for an account
      operationId: getAccountRisk
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
          example: user_42
      responses:
        '200':
          description: Risk assessment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/challenge:
    post:
      summary: Request an OTP challenge
      operationId: createChallenge
      description: |
        Generate a 6-digit OTP for step-up verification. The code expires
        after 5 minutes and allows 3 verification attempts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeRequest'
      responses:
        '201':
          description: Challenge created
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              $ref: '#/components/headers/Retry-After'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/challenge/verify:
    post:
      summary: Verify an OTP code
      operationId: verifyChallenge
      description: |
        Verify the 6-digit OTP. After 3 failed attempts, the challenge is
        invalidated and a new one must be requested.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponse'
        '400':
          description: Invalid request or challenge expired/exhausted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                expired:
                  value:
                    error: challenge_expired
                    message: This challenge has expired
                exhausted:
                  value:
                    error: attempts_exhausted
                    message: Maximum verification attempts exceeded
                invalid_code:
                  value:
                    error: invalid_code
                    message: Incorrect code. 2 attempts remaining.
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Challenge not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'